+++ packages/mosaic/src/chart-builders/types.ts
@@ -1,45 +1,12 @@
-import type {Spec} from '@uwdata/mosaic-spec';
 import type {ComponentType} from 'react';
-import type {VgPlotChartType} from '../chart-types/base-types';
+import type {
+  ChartTypeDefinition,
+  ChartBuilderField,
+  ChartBuilderColumn,
+} from '../chart-types/base-types';
 
-/**
- * Describes a field selector in a chart builder UI
- */
-export interface ChartBuilderField {
-  /** Unique key for this field */
-  key: string;
-  /** Display label */
-  label: string;
-  /** Whether the field is required */
-  required?: boolean;
-  /** Filter columns by DuckDB type (e.g. 'INTEGER', 'VARCHAR', 'DOUBLE') */
-  types?: string[];
-  /** Optional helper text for AI or custom UIs */
-  description?: string;
-}
-
-/**
- * Shared chart-type definition used by both the chart-builder UI and
- * assistant-driven chart creation.
- */
-export interface ChartTypeDefinition<TSettings = any> {
-  /** Unique identifier */
-  id: VgPlotChartType;
-  /** Short human-friendly name used in chart-type grids and prompts */
-  label?: string;
-  /** Short description of what this builder creates */
-  description: string;
-  /** Field selectors the user must fill in */
-  fields: ChartBuilderField[];
-  /** Generate a Mosaic spec from table name and selected field values */
-  createSpec: (tableName: string, values: TSettings) => Spec;
-  /** Generate a chart title from selected field values */
-  buildTitle?: (fieldValues: Record<string, string>) => string;
-  /** Optional availability override for a given table schema */
-  isAvailable?: (columns: ChartBuilderColumn[]) => boolean;
-  /** Optional extra assistant-facing description */
-  aiDescription?: string;
-}
+// Re-export for backward compatibility
+export type {ChartTypeDefinition, ChartBuilderField, ChartBuilderColumn};
 
 /**
  * Describes a chart builder template that generates Mosaic JSON specs
@@ -48,14 +15,6 @@ export interface ChartTypeDefinition<TSettings = any> {
 export interface ChartBuilderTemplate extends ChartTypeDefinition {
   /** Icon component */
   icon: ComponentType<{className?: string}>;
-}
-
-/**
- * Column info passed to chart builder UI
- */
-export interface ChartBuilderColumn {
-  name: string;
-  type: string;
 }
 
 /**
