+++ packages/mosaic/src/chart-types/base-types.ts
@@ -3,6 +3,8 @@
  * Separated to avoid circular dependencies.
  */
 
+import type {Spec} from '@uwdata/mosaic-spec';
+
 /**
  * Supported chart type identifiers
  */
@@ -15,3 +17,50 @@ export type VgPlotChartType =
   | 'heatmap'
   | 'box-plot'
   | 'custom-spec';
+
+/**
+ * Describes a field selector in a chart builder UI
+ */
+export interface ChartBuilderField {
+  /** Unique key for this field */
+  key: string;
+  /** Display label */
+  label: string;
+  /** Whether the field is required */
+  required?: boolean;
+  /** Filter columns by DuckDB type (e.g. 'INTEGER', 'VARCHAR', 'DOUBLE') */
+  types?: string[];
+  /** Optional helper text for AI or custom UIs */
+  description?: string;
+}
+
+/**
+ * Column info passed to chart builder UI
+ */
+export interface ChartBuilderColumn {
+  name: string;
+  type: string;
+}
+
+/**
+ * Shared chart-type definition used by both the chart-builder UI and
+ * assistant-driven chart creation.
+ */
+export interface ChartTypeDefinition<TSettings = any> {
+  /** Unique identifier */
+  id: VgPlotChartType;
+  /** Short human-friendly name used in chart-type grids and prompts */
+  label?: string;
+  /** Short description of what this builder creates */
+  description: string;
+  /** Field selectors the user must fill in */
+  fields: ChartBuilderField[];
+  /** Generate a Mosaic spec from table name and selected field values */
+  createSpec: (tableName: string, values: TSettings) => Spec;
+  /** Generate a chart title from selected field values */
+  buildTitle?: (fieldValues: Record<string, string>) => string;
+  /** Optional availability override for a given table schema */
+  isAvailable?: (columns: ChartBuilderColumn[]) => boolean;
+  /** Optional extra assistant-facing description */
+  aiDescription?: string;
+}
