+++ packages/mosaic/src/chart-types/box-plot/definition.ts
@@ -1,5 +1,5 @@
 import type {Spec} from '@uwdata/mosaic-spec';
-import type {ChartTypeDefinition} from '../../chart-builders/types';
+import type {ChartTypeDefinition} from '../base-types';
 import type {BoxPlotChartSettings} from './schema';
 import {
   buildDefaultChartTitle,
