+++ packages/mosaic/src/chart-types/ecdf/definition.ts
@@ -1,5 +1,5 @@
 import type {Spec} from '@uwdata/mosaic-spec';
-import type {ChartTypeDefinition} from '../../chart-builders/types';
+import type {ChartTypeDefinition} from '../base-types';
 import type {EcdfChartSettings} from './schema';
 import {
   buildDefaultChartTitle,
